fix(dsv4/compressor): align x dtype to fused weight in F.linear
fa0f33447 dropped the `x.float()` and the `.float()` weight casts on the
assumption that ckpt-loaded ``wkv`` / ``wgate`` are already bf16. On load
paths where the weights come through as fp32 (e.g. the full-model scratch
loader on the dev host) the fused weight ends up fp32 while ``x`` is bf16,
so ``F.linear(x, self._wkv_wgate_fused)`` raises

    RuntimeError: expected mat1 and mat2 to have the same dtype, but got:
    c10::BFloat16 != float

Align ``x`` to the fused weight's actual dtype. When the weight is bf16
(the hot path fa0f33447 targets) this is a no-op and the Tensor-Core GEMM
path is preserved; when the weight is fp32 it recovers the pre-fa0f33447
behaviour so the call doesn't crash.

+++ rtp_llm/models_py/modules/dsv4/fp8/compressor.py
@@ -472,7 +472,9 @@ class CompressorFP8(nn.Module):
 
         device = x.device
         out_dim = (1 + self.overlap) * self.head_dim
-        fused_out = torch.nn.functional.linear(x, self._wkv_wgate_fused)
+        fused_out = torch.nn.functional.linear(
+            x.to(self._wkv_wgate_fused.dtype), self._wkv_wgate_fused
+        )
         kv, score = fused_out[..., :out_dim], fused_out[..., out_dim:]
 
         cp_ctx = self._cp_ctx
@@ -507,7 +509,9 @@ class CompressorFP8(nn.Module):
 
         device = x.device
         out_dim = (1 + self.overlap) * self.head_dim
-        fused_out = torch.nn.functional.linear(x, self._wkv_wgate_fused)
+        fused_out = torch.nn.functional.linear(
+            x.to(self._wkv_wgate_fused.dtype), self._wkv_wgate_fused
+        )
         kv, score = fused_out[..., :out_dim], fused_out[..., out_dim:]
 
         kv_flat = kv.reshape(bsz, -1).contiguous()
